# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "zhuzhao——Go 模块化单体 IAM + 工单系统（三层鉴权 + 资源级工单）",
        "title": "zhuzhao API",
        "contact": {},
        "version": "1.0"
    },
    "paths": {
...
        "/api/v1/ticket-templates": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ticket"
                ],
                "summary": "工单模板列表",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            },
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ticket-admin"
                ],
                "summary": "新建工单模板",
                "parameters": [
                    {
                        "description": "模板定义",
                        "name": "req",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_tracerbiubiubiu_zhuzhao_internal_model.CreateTicketTemplateRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/ticket-templates/delete": {
            "post": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ticket-admin"
                ],
                "summary": "删除工单模板",
                "parameters": [
                    {
                        "type": "string",
                        "description": "模板编码",
                        "name": "code",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/ticket-templates/update": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ticket-admin"
                ],
                "summary": "更新工单模板（patch；code/type/org 不可改）",
                "parameters": [
                    {
                        "type": "string",
                        "description": "模板编码",
                        "name": "code",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "更新内容",
                        "name": "req",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_tracerbiubiubiu_zhuzhao_internal_model.UpdateTicketTemplateRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/ticket-templates/{code}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ticket"
                ],
                "summary": "工单模板详情",
                "parameters": [
                    {
                        "type": "string",
                        "description": "模板编码",
                        "name": "code",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/ticket-types": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ticket"
                ],
                "summary": "工单类型列表",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            },
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ticket-admin"
                ],
                "summary": "新建工单类型",
                "parameters": [
                    {
                        "description": "类型定义",
                        "name": "req",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_tracerbiubiubiu_zhuzhao_internal_model.CreateTicketTypeRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/ticket-types/delete": {
            "post": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ticket-admin"
                ],
                "summary": "删除工单类型（有工单禁删 → 409）",
                "parameters": [
                    {
                        "type": "string",
                        "description": "类型编码",
                        "name": "code",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/ticket-types/fields/replace": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ticket-admin"
                ],
                "summary": "全量替换类型字段集",
                "parameters": [
                    {
                        "type": "string",
                        "description": "类型编码",
                        "name": "code",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "字段集",
                        "name": "req",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_tracerbiubiubiu_zhuzhao_internal_model.ReplaceTypeFieldsRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/ticket-types/update": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ticket-admin"
                ],
                "summary": "更新工单类型（patch；code 不可改）",
                "parameters": [
                    {
                        "type": "string",
                        "description": "类型编码",
                        "name": "code",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "更新内容",
                        "name": "req",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_tracerbiubiubiu_zhuzhao_internal_model.UpdateTicketTypeRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/ticket-types/{code}/fields": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ticket"
                ],
                "summary": "工单类型字段定义",
                "parameters": [
                    {
                        "type": "string",
                        "description": "工单类型编码",
                        "name": "code",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/tickets": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ticket"
                ],
                "summary": "工单列表",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "页码",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "每页条数",
                        "name": "page_size",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "工单类型",
                        "name": "type_code",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "工单状态",
                        "name": "status",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "处理人过滤，仅支持 me（当前用户——工作台待办/已办卡数据源）",
                        "name": "assignee",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            },
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ticket"
                ],
                "summary": "创建工单",
                "parameters": [
                    {
                        "description": "创建工单请求",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_tracerbiubiubiu_zhuzhao_internal_model.CreateTicketRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/tickets/assign": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ticket"
                ],
                "summary": "分派工单",
                "parameters": [
                    {
                        "description": "分派工单请求",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_tracerbiubiubiu_zhuzhao_internal_model.AssignTicketRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/tickets/close": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ticket"
                ],
                "summary": "关闭工单",
                "parameters": [
                    {
                        "description": "关闭工单请求",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_tracerbiubiubiu_zhuzhao_internal_model.CloseTicketRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/tickets/comments": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ticket"
                ],
                "summary": "创建工单回复",
                "parameters": [
                    {
                        "description": "创建回复请求",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_tracerbiubiubiu_zhuzhao_internal_model.CreateCommentRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/tickets/delete": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ticket"
                ],
                "summary": "删除工单",
                "parameters": [
                    {
                        "description": "删除工单请求",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_tracerbiubiubiu_zhuzhao_internal_model.DeleteTicketRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "github_com_tracerbiubiubiu_zhuzhao_internal_model.CreateTicketTemplateRequest": {
            "type": "object",
            "required": [
                "code",
                "name",
                "org_id",
                "type_code"
            ],
            "properties": {
                "code": {
                    "type": "string",
                    "maxLength": 50
                },
                "default_fields": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "default_priority": {
                    "type": "integer"
                },
                "default_sla_minutes": {
                    "type": "integer"
                },
                "name": {
                    "type": "string",
                    "maxLength": 200
                },
                "org_id": {
                    "type": "string",
                    "example": "0"
                },
                "type_code": {
                    "type": "string",
                    "maxLength": 50
                }
            }
        },
...
        "github_com_tracerbiubiubiu_zhuzhao_internal_model.DeleteTicketRequest": {
            "type": "object",
            "required": [
                "id"
            ],
            "properties": {
                "id": {
                    "type": "string",
                    "example": "0"
                }
            }
        },
...
        "response.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "message": {
                    "type": "string"
                },
                "request_id": {
                    "type": "string"
                }
            }
        }
    }
}